+++ DebugMode/release-notes.md
@@ -2,7 +2,7 @@
 
 # Release notes
 ## 1.12.3
-Released 22 November 2020.
+Released 21 November 2020.
 
 * Improved translations. Thanks to PanPan-p (added Turkish)!